# server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: backend
  description: x-tutor APIs
basePath: /api
...
consumes:
  - application/json
produces:
  - application/json
definitions:
...
  newClass:
    type: object
    required:
    - tutor_id
    - student_id
    - subject
    - sessions
    - hourly_rate
    properties:
      sessions:
        type: array
        items:
          type: object
      tutor_id:
        type: string
      student_id:
        type: string
      subject:
        type: string 
      hourly_rate:
        type: number
  userInfo:
    type: object
    properties:
      academic_level_name:
        type: string
      # account_number:
      #   type: string
      address:
        type: string
      # bank_name:
      #   type: string 
      date_of_birth:
        type: string
      first_name:
        type: string
      gender_name:
        type: string
      last_name:
        type: string
      # country_name:
      #   type: string
      # language_name:
      #   type: string
      # city_name:
      #   type: string
      nationality_name:
        type: string
      # no_passport:
      #   type: string
      postal_code:
        type: string
      # profile_picture:
      #   type: string
      payment_method:
        type: string
      paypal_email:
        type: string
      # phone_code:
      #   type: string
      phone_number:
        type: integer
...
paths:
...
  /user:
    put:
      tags:
        - user
      description:
        - get users profile
      parameters:
        - account:
          in: body
          require: true
          description: update info user
          schema: 
            $ref: "#/definitions/userInfo"    
      respond:
        200: get all users      
...
  /class:
    post:
      tags:
        - class
      description: book new class
      parameters:
       - name: info
         in: body
         require: true
         description: infomation of new class booked by student
         schema: 
           $ref: "#/definitions/newClass"
      respond:
        - 200:
          description: Class success
        - 405:
          description: Invalid info
...